# Force-inline _modify accessors to work around a performance issue