@lorentey 2022-08-31
It turns out we did not have any benchmarks exercising this accessor. The improvement is pretty sizable:

<img width="1280" alt="01 Deque Int  modify through subscript (discontiguous)" src="https://user-images.githubusercontent.com/608696/187792746-5b96a26b-e17a-4aae-b061-134ce199dc28.png">
<img width="1280" alt="02 Deque Int  modify through subscript (contiguous)" src="https://user-images.githubusercontent.com/608696/187792752-0cefe2bd-2908-4a78-9d6c-688e051f4b35.png">

`malloc` is quite fast though! The extra 10ns does dominate this benchmark, but in higher-level code it can often be swamped by the actual work.