@Lukasa 2022-09-01
@lorentey Is that a 10ns slowdown? The y axis is logarithmic, so isn't this a 10x improvement? (Eyeballing logarithmic scales is hard but we appear to be about 1/3 of the way through the 10ns to 100ns branch in the "before" which would be about 20ns, as compared to also about 1/3 of the way through the 1ns to 10ns branch in the "after", which would be about 2ns.)

Regardless, I agree the gain is great. I also think this is a best-case scenario for malloc, as the allocator is uncontended. I'd be curious if we can run the benchmark concurrently in multiple threads. That should slow it down due to increased synchronization overhead between threads and produce an even more marked gain.

(I realise now that the benchmark tool is OSS, so feel free to tell me to buzz off and get my own data :wink:)