@lorentey 2022-09-01
Ah good point, I haven't considered allocations getting swamped. Is there typically no thread-local fast path?

The benchmarking facility is very intentionally single-threaded (benchmarking concurrent code well is tricky), but it wouldn't be too hard to introduce some artificial background task as an experiment. I'm not interested enough to do it myself right now though. :-P