@lorentey 2022-09-01
Yes you are right, the absolute values are closer to 2ns vs 20ns, so the measured cost for malloc/free is 18ns, not 10ns.