@Lukasa 2022-09-02
> Ah good point, I haven't considered allocations getting swamped. Is there typically no thread-local fast path?

To bring us home, the answer is "sometimes, it depends on the allocator". I can't speak to the Darwin allocator which I don't know well enough, but glibc _does_ have a thread-local fast path at the cost of massively increasing memory fragmentation, while allocators like jemalloc or tcmalloc will have a thread-local fast path that occasionally attempts to steal memory from across threads, causing coordination hits. In general, adding tcmalloc to a multithreaded program on Linux will greatly reduce its memory usage at the cost of mildly increasing its CPU usage.