@Catfish-Man 2022-09-02
Darwin has a lockless fast path, but it's disabled in daemons for fragmentation reasons